+++ packages/flame/lib/components.dart
@@ -16,7 +16,6 @@ export 'src/components/mixins/hoverable.dart';
 export 'src/components/mixins/keyboard_handler.dart';
 export 'src/components/mixins/parent_is_a.dart';
 export 'src/components/mixins/single_child_particle.dart';
-export 'src/components/mixins/tappable.dart';
 export 'src/components/nine_tile_box_component.dart';
 export 'src/components/parallax_component.dart';
 export 'src/components/particle_component.dart';
@@ -31,6 +30,7 @@ export 'src/components/sprite_group_component.dart';
 export 'src/components/text_box_component.dart';
 export 'src/components/text_component.dart';
 export 'src/components/timer_component.dart';
+export 'src/events/component_mixins/tappable.dart';
 export 'src/extensions/vector2.dart';
 export 'src/geometry/circle_component.dart';
 export 'src/geometry/polygon_component.dart';
